Extract out a 'No. of Player' field
